% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAL.R
\name{getSwiftKeyRandomSample}
\alias{getSwiftKeyRandomSample}
\title{Get a random sampling of swiftkey data}
\usage{
getSwiftKeyRandomSample(nrows = 100, locale = "en_US",
  textType = "twitter", seedDate = lubridate::today())
}
\arguments{
\item{nrows}{number of rows to read. Set to -1 for all lines.}

\item{locale}{which locale to use, can be any of those present. Defaults to "en_US"}

\item{textType}{which file type to use, can be any of those present. Defaults to "twitter"' @param locale which locale to be using}

\item{seedDate}{seed to use, defaults to today's date.}
}
\value{
a character vector of randomly chosen text
}
\description{
Gets a random set of rows given a seed. That seed defaults to today's date
but can be set to be any date. Those data are then cached according to
the global settings. If the same seed and settings are run multiple times,
this program reads from cached file instead of regenerating those
samples as those files are quite large.
}

