% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWords.R
\name{getWords}
\alias{getWords}
\title{Get vector of words from character vectors}
\usage{
getWords(x)
}
\arguments{
\item{x}{a character vector containing text}
}
\value{
a character vector containing words
}
\description{
This function splits words up on spaces and deletes punctuation
separating words (contractions still contain apostrophes).
}
\details{
These words are filtered so that
\itemize {
  \item{there are no empty ones.}
}
}
\examples{
getWords('foo bar baz.')
##
}

