% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAL.R
\name{readSwiftKeyLines}
\alias{readSwiftKeyLines}
\title{Read Swiftkey Lines from File}
\usage{
readSwiftKeyLines(nrows = 100, locale = "en_US", textType = "twitter")
}
\arguments{
\item{nrows}{number of rows to read. Set to -1 for all lines.}

\item{locale}{which locale to use, can be any of those present. Defaults to "en_US"}

\item{textType}{which file type to use, can be any of those present. Defaults to "twitter"}
}
\value{
a character vector containing the lines of text
}
\description{
Reads lines from a file of swiftkey data, given the provided package settings.
}

