% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAL.R
\name{readSwiftKeyLinesPath}
\alias{readSwiftKeyLinesPath}
\title{Read Swiftkey Lines from File given Path}
\usage{
readSwiftKeyLinesPath(x, blocksize = 100, blocks = 1)
}
\arguments{
\item{x}{path to file or open connection}

\item{blocksize}{number of lines to read per block}

\item{blocks}{total number of blocks}
}
\value{
character vector of lines of swiftkey data
}
\description{
This reads lines from a file of swiftkey data, given its file path.
It can do so using one (default) or more blocks, which are concatenated.
}
\examples{
## For simple run:
readSwiftKeyLines('/path/to/file', blocksize = 100, blocks = 1)
## This is equivalent to:
readSwiftKeyLines('/path/to/file', blocksize = 25, blocks = 4)
}

